Extracted the TestObserver as TestListener in its own module
